Classes' structure implemented.
For now I managed to write structure of the WebPage, Category, Topic and
Word classes (and their abstract parents in ParsingClasses.py). However,
their functioning has not been tested, so I am not sure whether it will work.
For now, I will focus on implementation of parse_data() methods for each of
these classes. Then I will test everything else.

Because it is not a main goal of the project (And actually I should not spend
too much time on this), I won't write any tests in Python.