Implemented parse_data for WebPage, fixed saving/reading process.
I made lots of alterations, which is not really good for one commit. Anyway...

1. Implemented parse_data method for WebPage class. For this also has been written
additional method get_categories.

2. In the main.py written create_and_save_web_page and load_web_page functions, that
act as a setup actions for the program. During their development I had to change type
of the parameter in the save_data and read_data methods to BinaryIO.

3. Also implemented to_dict method for all the classes. It allows to properly save data
of the class (in particular, save dict the way I want to) without the usage of actual names
of the class' variables (I don't need underscores in "_text", "_link" etc., because you cannot
pass them to the constructor later).