Implemnted conversion .dat into .csv. Added description for WebPage method.
Conversion is not implemented fully, it doesn't take into account
all the data for Word (like definitions or wordtype). This is only
a draft, but it is for now enough for me, because right now I can
get words sorted by topics and categories.

Also, I added description to parse_topics_for_specific_category method
in WebPage.